Implement cross-platform native functionality:
- Add DPI handling for Windows.
- Introduce multi-monitor support for macOS, Linux, and Windows.
- Add cross-platform notification capabilities.
- Implement WebView2 custom scheme handling and initial settings for Windows.
- Update CMakeLists to include new platform-specific source files.
- Refactor macOS delegate structure for modularity.
