+++ src/InfiniFrame.NativeBridge/Native/Platform/Mac/Delegates/AppDelegate.mm
