+++ src/InfiniFrame.NativeBridge/Native/Platform/Mac/Delegates/NavigationDelegate.mm
