+++ src/InfiniFrame.NativeBridge/Native/Platform/Mac/Delegates/UiDelegate.h
