+++ src/InfiniFrame.NativeBridge/Native/Platform/Mac/Delegates/UiDelegate.mm
