+++ src/InfiniFrame.NativeBridge/Native/Platform/Mac/Delegates/UrlSchemeHandler.h
