+++ src/InfiniFrame.NativeBridge/Native/Platform/Mac/Delegates/WindowDelegate.h
