+++ src/InfiniFrame.NativeBridge/Native/Platform/Windows/Core/WindowEvents.Win32.cpp
@@ -1,61 +1,12 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
-#include <ShellScalingApi.h>
-
 #include "Platform/Windows/Window.Win32.Internal.h"
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
-BOOL MonitorEnum(const HMONITOR monitor, HDC, LPRECT, const LPARAM arg) {
-    auto callback = reinterpret_cast<GetAllMonitorsCallback>(arg);
-    UINT dpiX, dpiY;
-    MONITORINFO info = {};
-    info.cbSize = sizeof(MONITORINFO);
-    GetMonitorInfo(monitor, &info);
-    GetDpiForMonitor(monitor, MDT_EFFECTIVE_DPI, &dpiX, &dpiY);
-    Monitor props = {};
-    props.monitor.x = info.rcMonitor.left;
-    props.monitor.y = info.rcMonitor.top;
-    props.monitor.width = info.rcMonitor.right - info.rcMonitor.left;
-    props.monitor.height = info.rcMonitor.bottom - info.rcMonitor.top;
-    props.work.x = info.rcWork.left;
-    props.work.y = info.rcWork.top;
-    props.work.width = info.rcWork.right - info.rcWork.left;
-    props.work.height = info.rcWork.bottom - info.rcWork.top;
-    props.scale = dpiY / 96.0;
-    return callback(&props) ? TRUE : FALSE;
-}
-
-void InfiniFrameWindow::ShowNotification(AutoString title, AutoString body) {
-    std::wstring wideTitle = ToUTF16String(title);
-    std::wstring wideBody = ToUTF16String(body);
-    if (m_impl->_notificationsEnabled && WinToast::isCompatible()) {
-        WinToastTemplate toast =
-            WinToastTemplate(WinToastTemplate::ImageAndText02);
-        toast.setTextField(wideTitle.c_str(), WinToastTemplate::FirstLine);
-        toast.setTextField(wideBody.c_str(), WinToastTemplate::SecondLine);
-        if (!m_impl->_iconFileName.empty())
-            toast.setImagePath(m_impl->_iconFileName);
-        WinToast::instance()->showToast(toast, m_impl->_toastHandler.get());
-    }
-}
-
-void InfiniFrameWindow::GetAllMonitors(GetAllMonitorsCallback callback) const {
-    if (callback) {
-        EnumDisplayMonitors(
-            nullptr, nullptr, reinterpret_cast<MONITORENUMPROC>(MonitorEnum), reinterpret_cast<LPARAM>(callback)
-        );
-    }
-}
-
 InfiniFrameDialog* InfiniFrameWindow::GetDialog() const {
     return m_impl->_dialog.get();
-}
-
-void InfiniFrameWindow::AddCustomSchemeName(const AutoStringConst scheme) {
-    if (scheme)
-        m_impl->_customSchemeNames.emplace_back(ToUTF16String(const_cast<AutoString>(scheme)));
 }
 
 void InfiniFrameWindow::SetClosingCallback(const ClosingCallback callback) {
