+++ src/InfiniFrame.NativeBridge/Native/Platform/Windows/Dpi.Win32.cpp
@@ -1,0 +1,10 @@
+// ---------------------------------------------------------------------------------------------------------------------
+// Imports
+// ---------------------------------------------------------------------------------------------------------------------
+#include "Platform/Windows/Window.Win32.Internal.h"
+// ---------------------------------------------------------------------------------------------------------------------
+// Code
+// ---------------------------------------------------------------------------------------------------------------------
+unsigned int InfiniFrameWindow::GetScreenDpi() const {
+    return GetDpiForWindow(m_impl->_hWnd);
+}
