+++ src/InfiniFrame.NativeBridge/Native/Platform/Windows/WebView/WebView2Attach.Win32.cpp
@@ -96,54 +96,7 @@ void InfiniFrameWindow::AttachWebView() {
     if (startupString.length() > 0)
         options->put_AdditionalBrowserArguments(startupString.c_str());
 
-    bool requiresAppSchemeRegistration = std::any_of(
-        m_impl->_customSchemeNames.begin(), m_impl->_customSchemeNames.end(),
-        [](const std::wstring& schemeName) { return _wcsicmp(schemeName.c_str(), L"app") == 0; }
-    );
-    bool appSchemeRegistrationSupported = false;
-
-    // Register custom schemes with WebView2 so top-level navigations like app://... are allowed.
-    if (!m_impl->_customSchemeNames.empty()) {
-        wil::com_ptr<ICoreWebView2EnvironmentOptions4> options4;
-        if (SUCCEEDED(options->QueryInterface(IID_PPV_ARGS(&options4))) && options4) {
-            appSchemeRegistrationSupported = true;
-            std::vector<wil::com_ptr<ICoreWebView2CustomSchemeRegistration>> registrations;
-            registrations.reserve(m_impl->_customSchemeNames.size());
-
-            for (const auto& schemeName : m_impl->_customSchemeNames) {
-                auto registration = Microsoft::WRL::Make<CoreWebView2CustomSchemeRegistration>(schemeName.c_str());
-                if (!registration)
-                    continue;
-
-                // Only the embedded-assets scheme uses app://localhost/... and should be
-                // treated as secure with an authority component.
-                if (_wcsicmp(schemeName.c_str(), L"app") == 0) {
-                    registration->put_HasAuthorityComponent(TRUE);
-                    registration->put_TreatAsSecure(TRUE);
-                }
-                registrations.emplace_back(registration);
-            }
-
-            if (!registrations.empty()) {
-                std::vector<ICoreWebView2CustomSchemeRegistration*> rawRegistrations;
-                rawRegistrations.reserve(registrations.size());
-                for (auto& registration : registrations)
-                    rawRegistrations.emplace_back(registration.get());
-
-                options4->SetCustomSchemeRegistrations(
-                    static_cast<UINT32>(rawRegistrations.size()), rawRegistrations.data()
-                );
-            }
-        }
-    }
-
-    if (requiresAppSchemeRegistration && !appSchemeRegistrationSupported) {
-        MessageBox(
-            m_impl->_hWnd,
-            L"This app requires WebView2 custom scheme registration for app://localhost/. Please update "
-            L"WebView2 Runtime to a version that supports ICoreWebView2EnvironmentOptions4.",
-            L"WebView2 Runtime Too Old", MB_OK | MB_ICONERROR
-        );
+    if (!RegisterCustomSchemesOnOptions(options.Get())) {
         m_impl->_isWebView2Initializing = false;
         return;
     }
@@ -249,15 +202,9 @@ void InfiniFrameWindow::AttachWebView() {
                             };
                             auto nav = std::make_shared<NavigateOnce>(NavigateOnce{this});
 
-                            wil::com_ptr<ICoreWebView2Settings> settings;
-                            HRESULT settingsResult = m_impl->_webviewWindow->get_Settings(&settings);
-                            if (FAILED(settingsResult) || !settings) {
-                                return FAILED(settingsResult) ? settingsResult : E_FAIL;
-                            }
-                            settings->put_AreHostObjectsAllowed(TRUE);
-                            settings->put_IsScriptEnabled(TRUE);
-                            settings->put_AreDefaultScriptDialogsEnabled(TRUE);
-                            settings->put_IsWebMessageEnabled(TRUE);
+                            HRESULT settingsResult = ApplyInitialWebViewSettings();
+                            if (FAILED(settingsResult))
+                                return settingsResult;
 
                             EventRegistrationToken webMessageToken;
                             m_impl->_webviewWindow->add_WebMessageReceived(
@@ -283,125 +230,7 @@ void InfiniFrameWindow::AttachWebView() {
                             m_impl->_webMessageReceivedToken = webMessageToken;
                             m_impl->_hasWebMessageReceivedToken = true;
 
-                            EventRegistrationToken webResourceRequestedToken;
-                            auto webview23 = m_impl->_webviewWindow.try_query<ICoreWebView2_23>();
-                            if (webview23) {
-                                webview23->AddWebResourceRequestedFilterWithRequestSourceKinds(
-                                    L"*", COREWEBVIEW2_WEB_RESOURCE_CONTEXT_ALL,
-                                    COREWEBVIEW2_WEB_RESOURCE_REQUEST_SOURCE_KINDS_ALL
-                                );
-                            } else {
-                                m_impl->_webviewWindow->AddWebResourceRequestedFilter(
-                                    L"*", COREWEBVIEW2_WEB_RESOURCE_CONTEXT_ALL
-                                );
-                            }
-                            m_impl->_webviewWindow->add_WebResourceRequested(
-                                Callback<ICoreWebView2WebResourceRequestedEventHandler>(
-                                    [this](ICoreWebView2*, ICoreWebView2WebResourceRequestedEventArgs* args) {
-                                        if (m_impl->_isClosingOrClosed.load(std::memory_order_acquire))
-                                            return S_OK;
-
-                                        wil::com_ptr<ICoreWebView2WebResourceRequest> req;
-                                        if (FAILED(args->get_Request(&req)) || !req)
-                                            return S_OK;
-
-                                        wil::unique_cotaskmem_string uri;
-                                        req->get_Uri(&uri);
-                                        std::wstring uriString = uri.get();
-                                        wil::com_ptr<ICoreWebView2HttpRequestHeaders> requestHeaders;
-                                        std::wstring requestOrigin;
-                                        if (SUCCEEDED(req->get_Headers(&requestHeaders)) && requestHeaders) {
-                                            wil::unique_cotaskmem_string originHeaderValue;
-                                            if (SUCCEEDED(requestHeaders->GetHeader(L"Origin", &originHeaderValue)) &&
-                                                originHeaderValue.get() != nullptr &&
-                                                originHeaderValue.get()[0] != L'\0') {
-                                                requestOrigin = originHeaderValue.get();
-                                            }
-                                        }
-
-                                        if (uriString.find(L"/_framework/blazor.modules.json") != std::wstring::npos) {
-                                            static constexpr BYTE emptyModuleArray[] = {'[', ']'};
-                                            wil::com_ptr<IStream> dataStream;
-                                            dataStream.attach(
-                                                SHCreateMemStream(emptyModuleArray, sizeof(emptyModuleArray))
-                                            );
-                                            if (!dataStream)
-                                                return S_OK;
-
-                                            std::wstring responseHeaders = L"Content-Type: application/json";
-                                            responseHeaders += L"\r\nAccess-Control-Allow-Methods: GET, HEAD, OPTIONS";
-                                            responseHeaders += L"\r\nAccess-Control-Allow-Headers: *";
-                                            if (!requestOrigin.empty()) {
-                                                responseHeaders += L"\r\nAccess-Control-Allow-Origin: " + requestOrigin;
-                                                responseHeaders += L"\r\nAccess-Control-Allow-Credentials: true";
-                                                responseHeaders += L"\r\nVary: Origin";
-                                            } else {
-                                                responseHeaders += L"\r\nAccess-Control-Allow-Origin: *";
-                                            }
-
-                                            wil::com_ptr<ICoreWebView2WebResourceResponse> response;
-                                            m_impl->_webviewEnvironment->CreateWebResourceResponse(
-                                                dataStream.get(), 200, L"OK", responseHeaders.c_str(), &response
-                                            );
-                                            args->put_Response(response.get());
-                                            return S_OK;
-                                        }
-                                        size_t colonPos = uriString.find(L':', 0);
-                                        if (colonPos > 0) {
-                                            std::wstring scheme = uriString.substr(0, colonPos);
-                                            auto it = std::find(
-                                                m_impl->_customSchemeNames.begin(), m_impl->_customSchemeNames.end(),
-                                                scheme
-                                            );
-
-                                            if (it != m_impl->_customSchemeNames.end() &&
-                                                m_impl->_customSchemeCallback != nullptr) {
-                                                int numBytes;
-                                                AutoString contentType = nullptr;
-                                                wil::unique_cotaskmem dotNetResponse(m_impl->_customSchemeCallback(
-                                                    const_cast<AutoString>(uriString.c_str()), &numBytes, &contentType
-                                                ));
-                                                auto freeContentType =
-                                                    wil::scope_exit([&contentType] { CoTaskMemFree(contentType); });
-
-                                                if (dotNetResponse != nullptr && contentType != nullptr) {
-                                                    std::wstring contentTypeWS = contentType;
-
-                                                    wil::com_ptr<IStream> dataStream;
-                                                    dataStream.attach(SHCreateMemStream(
-                                                        reinterpret_cast<const BYTE*>(dotNetResponse.get()), numBytes
-                                                    ));
-                                                    if (!dataStream)
-                                                        return S_OK;
-                                                    wil::com_ptr<ICoreWebView2WebResourceResponse> response;
-                                                    std::wstring responseHeaders = L"Content-Type: " + contentTypeWS;
-                                                    responseHeaders +=
-                                                        L"\r\nAccess-Control-Allow-Methods: GET, HEAD, OPTIONS";
-                                                    responseHeaders += L"\r\nAccess-Control-Allow-Headers: *";
-                                                    if (!requestOrigin.empty()) {
-                                                        responseHeaders +=
-                                                            L"\r\nAccess-Control-Allow-Origin: " + requestOrigin;
-                                                        responseHeaders +=
-                                                            L"\r\nAccess-Control-Allow-Credentials: true";
-                                                        responseHeaders += L"\r\nVary: Origin";
-                                                    } else {
-                                                        responseHeaders += L"\r\nAccess-Control-Allow-Origin: *";
-                                                    }
-                                                    m_impl->_webviewEnvironment->CreateWebResourceResponse(
-                                                        dataStream.get(), 200, L"OK", responseHeaders.c_str(), &response
-                                                    );
-                                                    args->put_Response(response.get());
-                                                }
-                                            }
-                                        }
-
-                                        return S_OK;
-                                    }
-                                ).Get(),
-                                &webResourceRequestedToken
-                            );
-                            m_impl->_webResourceRequestedTokenForCustomScheme = webResourceRequestedToken;
-                            m_impl->_hasWebResourceRequestedToken = true;
+                            AttachCustomSchemeHandler();
 
                             EventRegistrationToken permissionRequestedToken;
                             m_impl->_webviewWindow->add_PermissionRequested(
@@ -419,17 +248,6 @@ void InfiniFrameWindow::AttachWebView() {
                             );
                             m_impl->_permissionRequestedToken = permissionRequestedToken;
                             m_impl->_hasPermissionRequestedToken = true;
-
-                            if (!m_impl->_contextMenuEnabled)
-                                SetContextMenuEnabled(false);
-                            if (!m_impl->_zoomEnabled)
-                                SetZoomEnabled(false);
-                            if (!m_impl->_devToolsEnabled)
-                                SetDevToolsEnabled(false);
-                            if (m_impl->_transparentEnabled)
-                                SetTransparentEnabled(true);
-                            if (m_impl->_zoom != 100)
-                                SetZoom(m_impl->_zoom);
 
                             HRESULT addScriptHr = m_impl->_webviewWindow->AddScriptToExecuteOnDocumentCreated(
                                 js_wide.c_str(),
